pyproject.toml
[project]
name = "test-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "chardet>=5.2.0",
    "django>=5.2,<5.3",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "ipython>=9.1.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
]
